requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
fa6-icons==0.0.2
fastcore==1.7.1
fastlite==0.0.9
fh-bootstrap==0.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Markdown==3.7
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.4.4
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite-fts4==1.0.3
sqlite-utils==3.37
sqlite_minutils==3.37.0.post1
starlette==0.38.2
tabulate==0.9.0
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
